requirements.txt
# BankEaseAI v2.0 依赖
# 更新日期: 2025-10-02

# ============================================
# 核心Web框架 (FastAPI + Next.js)
# ============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ============================================
# 数据验证和序列化
# ============================================
pydantic[email]>=2.5.0
email-validator>=2.1.0

# ============================================
# 认证和安全
# ============================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5.0.0  # Pin to 4.x for passlib compatibility

# ============================================
# 数据库 (SQLAlchemy ORM)
# ============================================
sqlalchemy>=2.0.0
alembic>=1.13.0

# ============================================
# PDF 处理
# ============================================
pdfplumber>=0.11.5
pdfminer.six>=20231228
pypdfium2>=4.29.0,<4.30.1

# ============================================
# AI 处理 (可选)
# ============================================
openai>=1.0.0
anthropic>=0.45.2

# ============================================
# 数据处理和导出
# ============================================
pandas>=1.3.0,<2.3.0
numpy>=1.21.0,<1.27.0
openpyxl>=3.0.0

# ============================================
# 实用工具
# ============================================
python-dotenv>=1.0.0
tqdm>=4.65.0
Pillow>=10.0.0

# ============================================
# HTTP 客户端
# ============================================
httpx>=0.25.0
requests>=2.31.0

# ============================================
# 开发工具 (可选)
# ============================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.7.0
